refactor: change path com.mobile to com.mobile.rural
