+++ android/app/BUCK
@@ -35,7 +35,7 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.mobile",
+    package = "com.mobile.rural",
 )
 
 android_resource(
